Fixed trader logig can tests.

+++ src/AR.Hft.Process/Domain/Assessment.cs
@@ -9,6 +9,6 @@ namespace AR.Hft.Process.Domain
     public class Assessment
     {
         public string Symbol { get; set; }
-        public int Recommendation { get; set; } 
+        public double Recommendation { get; set; } 
     }
 }
